% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_control.R
\name{check_sls_control_function}
\alias{check_sls_control_function}
\title{Check package dependencies inside a control function}
\usage{
check_sls_control_function(f, packages = "rsaber", name = NULL)
}
\description{
Check package dependencies inside a control function
}
\keyword{internal}
