% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_main.R
\name{cps_bound}
\alias{cps_bound}
\title{SABER for Sequential Lot Sampling}
\usage{
cps_bound(x, tidx, alpha = NULL, control = NULL, saber_fit = NULL)
}
\arguments{
\item{x}{Numeric; vector denoting whether unit is defect (x=1) or not (x=0)}

\item{tidx}{Numeric; the vector containing the time point}

\item{alpha}{(numeric) Setting the credibility level.}

\item{control}{object of class \code{\link[=saber_control]{saber_control()}}}

\item{saber_fit}{Object of class \code{saber_cps}}
}
\value{
Returns an object of class \code{saber_cps}
}
\description{
Perform estimation and inference using the SABER method. Applies
to continuous production (one-at-a-time) sampling.
}
\details{
Argument \code{control} should contain the following elements:
\itemize{
\item \code{prior}: A vector of length 2 denoting the starting prior.
\item \code{lambda}: The parameter for the forgetting function.
\item \code{forget}: Character string or function.
\item \code{frame}: The frame of reference; "total", "unsampled", or "accepted".
}
}
\examples{

control <- saber_control(
  prior = c(1, 1),
  lambda = 0.95,
  forget = "exponential",
  frame = "total"
)

x <- c(0, 1, 0, NA, NA, 0, NA)
tidx <- seq_along(x)

fit <- cps_bound(
  x = x,
  tidx = tidx,
  alpha = 0.05,
  control = control
)

fit

}
