% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_control.R
\name{cps_control}
\alias{cps_control}
\title{Construct CPS simulation control settings}
\usage{
cps_control(
  n_init,
  nSim = 1000,
  n_forecast,
  nSim_forecast = 100,
  forecast_type = c("estimated", "optimistic", "pessimistic")
)
}
\arguments{
\item{n_init}{Initial number of units sampled.}

\item{nSim}{Number of Monte Carlo simulations for defect-count estimation.}

\item{n_forecast}{Number of future units forecast at a time.}

\item{nSim_forecast}{Number of forecast simulation paths.}

\item{forecast_type}{Forecasting mode.}
}
\description{
Construct CPS simulation control settings
}
\examples{

cps_control( 
  n_init=100,
  nSim = 1000,
  n_forecast = 10,
  nSim_forecast = 100,
  forecast_type = "estimated"
)


}
