% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_main.R
\name{cps_prob_sample}
\alias{cps_prob_sample}
\title{Probability for sampling next unit in CPS}
\usage{
cps_prob_sample(Xsim_mc, tidx, preq, alpha, forecast_type, control)
}
\arguments{
\item{Xsim_mc}{(matrix) The MCMC for the current time step.}

\item{tidx}{(numeric) Vector of time points associated with each unit.}

\item{preq}{(numeric) The requirement on the defect rate.}

\item{alpha}{(numeric) Setting the credibility level.}

\item{forecast_type}{(string) Type of forecast, see details.}

\item{control}{object of class \code{\link[=saber_control]{saber_control()}}}
}
\value{
A numeric scalar giving the probability of sampling the next unit.
}
\description{
Computes a probability for sampling subsequent units in a continuous
production sampling framework.
}
\details{
Argument \code{forecast_type} can take values:
\itemize{
\item "optimistic" : assumes the next unit is not a defect
\item "pessimistic": assumes the next unit is a defect
\item "estimated": uses a weighted average of the optimistic and pessimistic probabilities
}
}
\examples{

control <- saber_control(
  prior = c(1, 1),
  lambda = 0.95,
  forget = "exponential",
  frame = "total"
)

x <- c(0, 1, 0, NA, NA)
tidx <- seq_along(x)

fit <- cps_bound(
  x = x,
  tidx = tidx,
  alpha = 0.05,
  control = control
)

cps_prob_sample(
  saber_fit = fit,
  pest = fit[["pest"]][nrow(fit)],
  preq = 0.10,
  alpha = 0.05,
  tidx_next = max(tidx) + 1,
  forecast_type = "estimated"
)


}
\seealso{
\code{\link[=cps_bound]{cps_bound()}}
}
