% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_main.R
\name{cps_sim_case}
\alias{cps_sim_case}
\title{Simulate a case for continuous production sampling}
\usage{
cps_sim_case(
  y,
  n,
  p,
  tidx,
  preq,
  alpha,
  control,
  cps_control,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{y}{(numeric) Vector of the true responses.}

\item{n}{(numeric) Vector denoting whether a unit was samples (n=1) or not (n=0), or not yet produced (n=NA).}

\item{p}{(numeric) Scalar or vector of the true defect rate.}

\item{tidx}{(numeric) Scalar or vector of the true defect rate.}

\item{preq}{(numeric) Requirement on defect rate.}

\item{alpha}{(numeric) Tail area for credibility level (\code{alpha = 0.05} means 95\% credibility).}

\item{control}{object of class \code{saber_control}.}

\item{cps_control}{object of class \code{cps_control}.}

\item{seed}{(numeric) Seed for the random number generator.}
}
\value{
A list with the following elements:
\itemize{
\item \code{saber_fit}: An object of class \code{saber_cps} containing the final fitted
SABER-CPS trajectory.
\item \code{y}: The true response for each unit (may have been supplied, may have been generated).
\item \code{x}: The observed response for each unit.
\item \code{Xsim_mc}: A matrix of Monte Carlo simulated defective-count contributions
for the CPS sequence.
\item \code{prob_sampled}: A numeric vector giving the sampling probability assigned
to each unit.
\item \code{control}: The \code{saber_control} object used for the SABER model.
\item \code{cps_control}: The \code{cps_control} object used for the CPS simulation.
}
}
\description{
Simulates a single run for continuous production sampling and applies
SABER-CPS. Used for assessing performance and parameterizing the forgetting
function.
}
\details{
Either \code{y} or \code{p} must be provided. The total number of units is determined by the length of \code{n}.
If provided, \code{y} is used instead of \code{p}. When providing \code{p}, it may be vectorized (e.g., to
simulate drift or step-changes).

Argument \code{control} should be the result of \code{\link[=saber_control]{saber_control()}}, setting the
parameterization for the SABER model. It must contain the following elements:
\itemize{
\item \code{prior}: A vector of length 2 denoting the starting prior.
\item \code{lambda}: The parameter for the forgetting function.
\item \code{forget}: Character string or function.
\item \code{frame}: The frame of reference; "total", "unsampled", or "accepted".
}

Argument \code{cps_control} should be the result of \code{\link[=cps_control]{cps_control()}}, setting the
parameterization for the simulation. It must contain the following elements:
\itemize{
\item \code{nSim}  : Number of Monte Carlo simulations for defect-count estimation.
\item \code{forecast_type}  : Forecasting mode.
}
}
\examples{

\dontrun{
control <- saber_control(
  prior = c(1, 1),
  lambda = 0.95,
  forget = "exponential",
  frame = "total"
)

cps_ctrl <- cps_control(
  nSim = 100,
  forecast_type = "estimated"
)

n_vector   <- rep( c(1,NA), c(20,80) )
unit_times <- cumsum(rexp(length(n_vector)))

sim <- cps_sim_case(
  p = 0.05,
  n = n_vector,
  tidx = unit_times,
  preq = 0.10,
  alpha = 0.05,
  control = control,
  cps_control = cps_ctrl,
  seed = 123
)

names(sim)
class(sim$saber_fit)
}

}
