% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinomial_dist.r
\name{dbetabinomial}
\alias{dbetabinomial}
\alias{pbetabinomial}
\alias{qbetabinomial}
\alias{rbetabinomial}
\title{The Beta-Binomial Distribution}
\usage{
dbetabinomial(y, Nn, a, b)

pbetabinomial(y, Nn, a, b)

qbetabinomial(p, Nn, a, b)

rbetabinomial(n, Nn, a, b)
}
\arguments{
\item{y}{Vector of quantiles}

\item{Nn}{number of observations}

\item{a}{alpha parameter}

\item{b}{beta parameter}

\item{p}{Vector of probabilities
\code{round} (default), \code{ceiling}, or \code{floor}. If}

\item{n}{Number of random variates to draw}
}
\value{
\code{dbetabinomial} gives the mass function (PMF),
\code{pbetabinomial} gives the distribution function (CDF),
\code{qbetabinomial} gives the quantile function, and
\code{rbetabinomial} generates random deviates.
}
\description{
Density, distribution function, quantile function and random
generation for the Beta-Binomial distribution.
}
\details{
The Beta-Binomial distribution has the mass function

\deqn{
p(x) = \binom{Nn}{y}\dfrac{B(y+a, Nn-y+b)}{B(a,b)}
}

where \eqn{B()} is the Beta function. The syntax \code{Nn} is used so that
\code{rbetabinomial} can use \code{n} for the number of random variates to draw.

The quantile function \code{qbetabinomial} uses an optimization which does not
necessarily return an integer. The final result is rounded to the nearest
integer. Investigations show this to be a very small rounding (on the order
of 1e-4 or less).

\code{rbetabinomial} does not use the usual inverse of the CDF. For numerical
simplicity and vectorization it uses the fact that the BetaBinomial is a
Binomial random variable with unknown probability. Variates are obtained by
sampling the Beta distribution with the appropriate prior, then sampling the
Binomial distribution with that Beta variate.
}
\examples{

dbetabinomial( 1, 10, 1, 5 )
dbetabinomial( 0:10, 10, 1, 5 )



dbetabinomial( 1, 10, 1, 5 )
dbetabinomial( 0:10, 10, 1, 5 )




qbetabinomial( 0.35, 10, 1, 5 )
qbetabinomial( seq(0,1,0.10), 10, 1, 5 )



rbetabinomial( 20, 10, 1, 5 )


}
