% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parameter_mesh}
\alias{parameter_mesh}
\title{Create a mesh of parameters}
\usage{
parameter_mesh(mesh, params)
}
\arguments{
\item{mesh}{(string) a character string denoting the type of meshing to be done.}

\item{params}{(list) a named list of parameters to be meshed.}
}
\value{
Returns a data frame.
}
\description{
Takes a list of input parameters and creates a mesh in one of
several manners.
}
\details{
Argument \code{mesh} has two options:
\itemize{
\item "crossing" will create all combinations (using \code{expand.grid})
\item "recycle" will cycle through the elements, repeating as needed to fill the grid.
Users should carefully inspect the resulting mesh if this option is used.
}
}
