% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_methods.r
\name{prop_n_freq}
\alias{prop_n_freq}
\title{Sample size estimation for reliability}
\usage{
prop_n_freq(
  x = 0,
  N = Inf,
  preq,
  alpha,
  mesh = "crossing",
  method = "wilson",
  maxiter = 100,
  preqdigits = 8
)
}
\arguments{
\item{x}{Numeric (integer); the number of events observed, defaults to 0}

\item{N}{Numeric (integer); Size of the population (can be \code{Inf} for infinite-population case)}

\item{preq}{Numeric (decimal); the requirement on the defect rate}

\item{alpha}{Numeric (decimal); complement of the confidence level}

\item{mesh}{character string identifying how to mesh the parameters, see \code{parameter_mesh}.}

\item{method}{Character; Use original Darby ("darby") formulation,
or Lauren Wilson's modification ("wilson").}

\item{maxiter}{Numeric (integer); maximum number of iterations to permit. See notes.}

\item{preqdigits}{Numeric (integer); number of digits to round preq, defaults to 8}
}
\value{
Returns a tibble with columns for N, preq, alpha, and smallest sample size.
}
\description{
Implementation of several a method of obtaining minimum sample size needed
to assert a specific reliability requirement (defined in terms of the defect
rate). Leverages Clopper-Pearson type upper bound on defect rate.
}
\details{
The method implemented assumes a Clopper-Pearson type (or "exact") one-sided
upper confidence bound on the defect rate, and sets the sample size such that
this upper bound will be less than or equal to the requirement.
}
\note{
The function will estimate an upper bound based on a zero-failure scheme. If
x>0 defects are allowed, this may not be a large enough upper bound. The
function uses a while loop to slowly increase the sample size, the \code{maxiter}
parameter is there to prevent infinite loops.
The \code{preqdigits} argument was added due to machine precision issues. For
example, a value of 0.05 may produce a different results to 1 - 0.95. This
value defaults to 8, which should be suitable for most cases.

prop_n_freq(
x = 0,
N = 100,
preq = 0.05,
alpha = c(0.1, 0.05),
)
}
