% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_methods.r
\name{prop_interval_bayes}
\alias{prop_interval_bayes}
\title{Bayesian point and interval estimation for proportions}
\usage{
prop_interval_bayes(
  x,
  n,
  N,
  alpha,
  prior = c(1, 1),
  mesh = "crossing",
  frame = "total"
)
}
\arguments{
\item{x}{Numeric (integer); the number of events observed}

\item{n}{Numeric (integer); the number of samples taken}

\item{N}{Numeric (integer); Size of the population (can be \code{Inf} for infinite-population case)}

\item{alpha}{numeric, complement of the confidence level}

\item{prior}{numeric vector of the Beta prior, defautls to \code{c(1,1)}.}

\item{mesh}{character string identifying how to mesh the parameters, see \code{parameter_mesh}.}

\item{frame}{character string identifying the frame of reference for finite populations,
currently accepted values are "total", "accepted", or "unsampled."}
}
\value{
Returns a list, or a tibble with columns for point and interval estimates.
For the finite-population case, this includes estimates of both the proportion
of events as well as the number of events in the population (\emph{including} the observed events).
}
\description{
Uses Bayesian approach to estimate the proportion for binary event data.
For the finite-population case, a superpopulation approach is used, leading to a
Beta-Binomial posterior distribution. The prior distribution defaults to a uniform
(non-informative) distribution.
}
\details{
Computes a "margin of error" as the larger of the differences
$\eqn{X_{1-\alpha}-X_{0.5}}$ and $\eqn{X_{0.5}-X_{\alpha}}$

For finite populations, the posterior distribution is Beta-Binomial with
population size N-n (that is, there are N-n untested units in the population).
There are several frames of reference which may be of interest here.
\itemize{
\item \code{frame="total"} corresponds to the entire population, including the
units that were tested, and those which are defective.
\item \code{frame="accepted"} corresponds to only accepted units. This will \emph{include} tested units
but will \emph{exclude} tested units that are defective.
\item \code{frame="unsampled"} corresponds to only untested units. This will exclude
all tested units, regardless of whether they are defective or not.
}

The "accepted" case relates to non-destructive (and non-degrading) testing,
while "unsampled" relates to destructive testing.
}
\note{
The Hypergeometric interval on p for finite population uses Bayesian estimation.
The default is a Beta(1,1) prior on p. An example the BetaBinomial is noninformative:
hist(  rsaber::qbetabinomial( runif(10000,0,1) , Nn=100, a=1, b=1)  )
}
\examples{
prop_interval_bayes(
  x = 0,
  n = c(10,20,30,40),
  N = 100,
  alpha = c(0.1, 0.05),
  prior = c(1,1)
)



}
