% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_methods.r
\name{prop_interval_freq}
\alias{prop_interval_freq}
\title{Frequentist point and interval estimation proportions}
\usage{
prop_interval_freq(x, n, N, alpha, mesh = "crossing", type = "wilson")
}
\arguments{
\item{x}{Numeric (integer); the number of events observed}

\item{n}{Numeric (integer); the number of samples taken}

\item{N}{Numeric (integer); Size of the population (can be \code{Inf} for infinite-population case)}

\item{alpha}{Numeric (decimal); complement of the confidence level}

\item{mesh}{character string identifying how to mesh the parameters, see \code{parameter_mesh}.}

\item{type}{Character; the type of bound, currently supported options are
"wilson" (default), "agresti-coull", "clopper-pearson", and "wald"}
}
\value{
Returns a list or tibble with columns for point and interval estimates.
}
\description{
Implements Frequentist approaches to estimate the proportion for binary
event data. Methods available are Wilson, Agresti-Coull, Clopper-Pearson, Jeffreys, and Wald.
Applies a finite-population correction factor (FPC) when argument \code{N} is non-infinite.
}
\details{
The argument \code{type} is converted to lowercase, and several abbreviations are permitted
(wils for wilson, agresti, ac, and a-c for agresti-coull).

The Wald interval tends to demonstrate poor coverage, particularly when p is close
to 0 or 1 and/or n is small. It is included for completeness, but should rarely be
the preferred method. The default method is the Wilson (score) interval.

For the Clopper-Pearson interval is obtained using the Beta representation. This implies
two different "priors". To resolve this for the estimate, we use the result from the
Jeffreys interval.
}
\examples{
prop_interval_freq(
  x = 0,
  n = c(10,20,30,40),
  N = 100,
  alpha = c(0.1, 0.05),
)

}
