% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_methods.r
\name{prop_n_bayes}
\alias{prop_n_bayes}
\title{Bayesian method to determine sample size for reliability requirement}
\usage{
prop_n_bayes(
  x,
  N,
  preq,
  alpha,
  prior = c(1, 1),
  mesh = "crossing",
  frame = "total",
  maxiter = 100,
  verbose = TRUE
)
}
\arguments{
\item{x}{Numeric (integer); the number of events observed}

\item{N}{Numeric (integer); Size of the population (can be \code{Inf} for infinite-population case)}

\item{preq}{Numeric (decimal); the requirement on the defect rate}

\item{alpha}{Numeric (decimal); complement of the confidence level}

\item{prior}{Numeric (float); value of \eqn{\alpha} and \eqn{\beta} in the prior distribution for \eqn{\pi}. Defautls to \code{c(1,1)}.}

\item{mesh}{character string identifying how to mesh the parameters, see \code{parameter_mesh}.}

\item{frame}{Character; the frame of reference for finite populations, currently accepted values are "total", "accepted", or "unsampled."}

\item{maxiter}{Numeric (integer); maximum number of iterations to permit. See notes.}

\item{verbose}{Logical; whether or not to print non-error messages.}
}
\value{
Returns a tibble with columns for N, preq, alpha, and smallest sample size.
}
\description{
Bayesian approach of obtaining minimum sample size needed to assert
a specific reliability requirement (defined in terms of the defect rate).
For the finite-population case, a superpopulation approach is used, leading to a
Beta-Binomial posterior distribution. The prior distribution defaults to a uniform
(non-informative) distribution.
}
\note{
The function will estimate an upper bound based on a zero-failure scheme. If
x>0 defects are allowed, this may not be a large enough upper bound. The
function uses a while loop to slowly increase the sample size, the \code{maxiter}
parameter is there to prevent infinite loops.
}
\examples{
prop_n_bayes(
  x = 0,
  N = 100,
  preq = 0.05,
  alpha = c(0.1, 0.05),
  prior = c(1,1)
)

}
