% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saber_general.R
\name{saber_posterior}
\alias{saber_posterior}
\title{Posterior parameters for SABER}
\usage{
saber_posterior(x, n, tidx, prior = c(1, 1), lambda, forget = "exponential")
}
\arguments{
\item{x}{Numeric vector giving the observed number of defectives for each
observation.}

\item{n}{Numeric vector giving the sample size or binomial denominator
associated with each observation.}

\item{tidx}{Numeric vector giving the time/order index for each observation.
Used to determine recency in the forgetting function.}

\item{prior}{Numeric vector of length 2 giving the beta prior parameters.}

\item{lambda}{Numeric scalar or vector giving the parameter for the forgetting function.}

\item{forget}{String; the type of forgetting function.}
}
\value{
A numeric vector of length 2 containing the posterior
beta parameters \code{c(a, b)}.
}
\description{
Obtain the posterior parameters for SABER
}
\examples{
x <- c(0, 1, 0, 2)
n <- c(10, 10, 10, 10)
tidx <- seq_along(x)

saber_posterior(
  x = x,
  n = n,
  tidx = tidx,
  prior = c(1, 1),
  lambda = 0.95,
  forget = "exponential"
)

saber_posterior(
  x = x,
  n = n,
  tidx = tidx,
  prior = c(1, 1),
  lambda = 2,
  forget = "fixed_window"
) 

}
\seealso{
\code{\link[=saber_weights]{saber_weights()}}
}
