% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saber_general.R
\name{saber_weights}
\alias{saber_weights}
\title{Compute SABER forgetting weights}
\usage{
saber_weights(n, tidx, lambda, forget = "exponential")
}
\arguments{
\item{n}{Numeric vector giving the sample size or binomial denominator
associated with each observation.}

\item{tidx}{Numeric vector giving the time/order index for each observation.
Used to determine recency in the forgetting function.}

\item{lambda}{Numeric scalar or vector giving the parameter(s) for the forgetting function.}

\item{forget}{Character string or function specifying the forgetting method.
Character options are \code{"exponential"}, \code{"fixed_window"},
\code{"total_samples"}, and \code{"none"}.}
}
\value{
Returns a numeric vector
}
\description{
Evaluates the phi function in SABER for weighting data.
}
\details{
Custom functions for \code{forget} should accept arguments \code{n}, \code{tidx}, and
\code{lambda}, and return a numeric weight vector with the same length as \code{n}.
}
\note{
The meaning of \code{n} and \code{lambda} changes depending on the forgetting function
and type of SABER model. For SABER-SLS (when running \code{\link[=sls_bound]{sls_bound()}}), \code{n}
denotes the sample size from each lot. For SABER-CPS (when running
\code{\link[=cps_bound]{cps_bound()}}), \code{n} is a vector with all 1's.
}
\examples{
n <- rep(10, 8)
tidx <- seq_along(n)

# Exponential forgetting: older observations receive smaller weights.
saber_weights(
  n = n,
  tidx = tidx,
  lambda = 0.90,
  forget = "exponential"
)

# Fixed-window forgetting: only observations within the window receive weight.
saber_weights(
  n = n,
  tidx = tidx,
  lambda = 3,
  forget = "fixed_window"
)

# No forgetting: all observations receive weight one.
# Can also be accomplished withforget= "exponential" and lambda=1
saber_weights(
  n = n,
  tidx = tidx,
  lambda = 1,
  forget = "none"
) 


}
