% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Set the number of decimal places for a value}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{(Numeric) the value to be set}

\item{k}{(numeric integer) the number of decimal places}
}
\value{
Returns a string value.
}
\description{
Forces a particular decimal representation of a numeric value. This
converts the result to character.
}
