% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_control.R
\name{validate_cps_control}
\alias{validate_cps_control}
\title{Validate CPS simulation control settings}
\usage{
validate_cps_control(control)
}
\description{
Validate CPS simulation control settings
}
\keyword{internal}
