% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saber_general.R
\name{validate_saber_control}
\alias{validate_saber_control}
\title{Validate SABER model control settings}
\usage{
validate_saber_control(control)
}
\description{
Validate SABER model control settings
}
\keyword{internal}
