% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_saber_fit.R
\name{validate_saber_fit}
\alias{validate_saber_fit}
\title{Validate a SABER fit object}
\usage{
validate_saber_fit(
  x,
  subclass = NULL,
  required_cols = NULL,
  require_control = TRUE
)
}
\description{
Validate a SABER fit object
}
\keyword{internal}
