% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_control.R
\name{validate_sls_control}
\alias{validate_sls_control}
\title{Validate SLS simulation/control settings}
\usage{
validate_sls_control(control)
}
\description{
Validate SLS simulation/control settings
}
\keyword{internal}
