% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sls_control.R
\name{validate_sls_packages}
\alias{validate_sls_packages}
\title{Validate package list for SLS parallel execution}
\usage{
validate_sls_packages(packages)
}
\description{
Validate package list for SLS parallel execution
}
\keyword{internal}
