% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saber_general.R
\name{saber_control}
\alias{saber_control}
\title{Construct SABER model control settings}
\usage{
saber_control(
  prior = c(1, 1),
  lambda = NULL,
  forget = c("exponential", "fixed_window", "total_samples", "none"),
  frame = c("total", "unsampled", "accepted")
)
}
\arguments{
\item{prior}{Numeric vector of length 2 giving beta prior parameters.}

\item{lambda}{Forgetting parameter.}

\item{forget}{Forgetting method or custom forgetting function.}

\item{frame}{Frame of reference for defect estimation.}
}
\description{
Construct SABER model control settings
}
\examples{
# Exponential forgetting with the total-population frame.
control <- saber_control(
  prior = c(1, 1),
  lambda = 0.95,
  forget = "exponential",
  frame = "total"
)

control

# No forgetting.
saber_control(
  prior = c(1, 1),
  forget = "none",
  frame = "total"
) 

}
